django_api/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
Cython==0.29.21
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-rq==2.3.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pipenv==2020.8.13
plotly==4.10.0
psycopg2-binary==2.8.6
pycosat==0.6.3
Pygments==2.7.1
PyJWT==1.7.1
python2==1.2
pytz==2020.1
redis==3.5.3
requests==2.24.0
retrying==1.3.3
rq==1.5.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
standardjson==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
webhooks==0.4.2
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.4.0

